pyproject.toml
[project]
name = "odpyx"
version = "0.1.0"
description = "Server emulator for a certain gacha game."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0",
    "requests>=2.32",
    "pycryptodome>=3.20",
    "msgspec>=0.18.6",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
